#103: add generic core_backends mixin with temporary alias
Extract TrustModelBackendMixin and PlanQueryCompiler wiring to
trusts.core_backends. Keep trusts.backends.TrustModelBackendMixin as
the exact same object and leave the historical backend path unchanged.
