+++ tests/runtests.py
@@ -21,6 +21,7 @@ KERNEL_SUITE = [
     'trusts.test_issue96',
     'trusts.test_issue98',
     'trusts.test_issue100',
+    'trusts.test_issue103',
 ]
 
 # C1 ran ``run_tests(['trusts'])``. Pair re-runs that full package except
@@ -49,6 +50,7 @@ PAIR_LEGACY_SUITE = [
     'trusts.test_issue92',
     'trusts.test_issue98',
     'trusts.test_issue100',
+    'trusts.test_issue103',
 ]
